requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0

# Optional: for AI-powered insights
anthropic>=0.40.0

